Partfs: add truncate support.
This commit adds file-truncation support to partfs mounts.
Trunctation doesn't actually affect anything on the underlying
storage file, but it does affect the presented size of the mount.

With the addition of truncation support, a partfs mount will look
like any file copied into it, and will resize itself accordingly.

Note that the actual storage will never exceed the value of
-o sizelimit, and also that trunctation doesn't affect the size
of the underlying file.