Finish adding in new partition-table features.
This commit finishes the changes to the build system, manpage,
and core utility.

As of this commit, partfs should support a new '-o partition' syntax
that can be used to pick a partition without needing to know its
offset/size. Additionally, a new -p/--print-partitions flag is added
that can be used to dump the partition table of the input file.